% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbR-package.R
\docType{data}
\name{Elevation}
\alias{Elevation}
\title{NWT-2023 Elevation Dataset}
\format{
rdata
}
\source{
USGS website: https://search.earthdata.nasa.gov/search/granules?projectId=2435656526Example usage of the Elevation dataset includes deriving slope and aspect

\code{
# Load the dataset (assuming you've already installed and loaded CarbR package)
data("Elevation")

# Derive slope and aspect using appropriate functions
slope_result <- calculate_slope(Elevation)
aspect_result <- calculate_aspect(Elevation)
}
}
\usage{
data(Elevation)
}
\description{
A dataset containing the Digital Elevation Model (DEM) in meters for Northwestern Territory, Canada for the year 2023.
Areas of higher values depict areas of higher elevation, and areas with low values depict areas with low elevation.
The data was obtained from the USGS website.
}
\references{
\url{https://github.com/p/ChinyereRuth/CarbR/}
}
\author{
Chinyere Ottah, School of Earth, Environment and Society, McMaster Univetsity
}
\keyword{dataset}
