% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbR-package.R
\docType{package}
\name{CarbR-package}
\alias{CarbR}
\alias{CarbR-package}
\title{CarbR: Deriving Slope and Aspect from DEM}
\description{
A package for deriving slope and aspect using Digital Elevation Models (DEMs) in R.

A package for deriving slope and aspect using Digital Elevation Models (DEMs) in R.

This package provides a workflow of package creation using R stidio. The package includes a function
to derive slope and aspect using DEM and also a documented for the function.
The DEM used for the study site is also included in the package.
}
\references{
\url{https://github.com/p/ChinyereRuth/CarbR/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChinyereRuth/CarbR}
  \item Report bugs at \url{https://github.com/ChinyereRuth/CarbR/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/ChinyereRuth/CarbR}
  \item Report bugs at \url{https://github.com/ChinyereRuth/CarbR/issues}
}

}
\author{
\strong{Maintainer}: Chinyere Ottah \email{chinyere.ottah@mcmaster.ca} (\href{https://orcid.org/0000-0002-2725-9448}{ORCID})


Chinyere Ottah, School of Earth, Environment and Society, McMaster Univetsity
}
\keyword{internal}
